Resolve "[meta] Denseの連立一次方程式ソルバを実装する"

Closes #374

See merge request ricos/monolish!240

+++ include/monolish_solver.hpp
@@ -69,7 +69,7 @@ public:
    * @brief set library option (to change library, monolish, cusolver, etc.)
    * @param[in] library number
    **/
-  void set_lib(double l) { lib = l; }
+  void set_lib(int l) { lib = l; }
 
   /**
    * @brief set tolerance (default:1.0e-8)
